==> experiment1/10.Combine Numerical and TF-IDF Features.ipynb <==
In this notebook we scale numerical features. 
- Numerical features should be scaled because they have different ranges and magnitudes, and scaling ensures that each feature contributes equally to the analysis, preventing features with larger ranges from dominating the results.
- TF-IDF vectors are already normalized, with values generally in the range [0, 1], representing term frequencies normalized by document frequencies. Scaling these again could distort their meaning.

### Get Data

from sklearn.preprocessing import StandardScaler
import pandas as pd

# Load the dataset
df = pd.read_csv('numeric_features_added_v1.csv')
tfidf_features_df = pd.read_csv('tfidf_features.csv')

### Check the range of the TF-IDF values in your tfidf_features_df

# Check the minimum and maximum values in the DataFrame
min_tfidf_value = tfidf_features_df.min().min()
max_tfidf_value = tfidf_features_df.max().max()

print(f"Minimum TF-IDF value: {min_tfidf_value}")
print(f"Maximum TF-IDF value: {max_tfidf_value}")

### Scaling numerical features

# Select numerical features from df
numerical_features = [
    'reading_time',
    'mistakes_dist_ratio',
    'polysyllabcount',
    'sentence_count',
    'difficult_words',
    'comma_count',
    'transitional_phrases_c',
    'text_dist_words_ratio'
]

# Create a DataFrame for numerical features
numerical_features_df = df[numerical_features]
numerical_features_df.head(1)

# Find the range of 'mistakes_dist_ratio' and 'text_dist_words_ratio' features
mistakes_dist_ratio_range = (df['mistakes_dist_ratio'].min(), df['mistakes_dist_ratio'].max())
text_dist_words_ratio_range = (df['text_dist_words_ratio'].min(), df['text_dist_words_ratio'].max())

print(f"Range of 'mistakes_dist_ratio': {mistakes_dist_ratio_range}")
print(f"Range of 'text_dist_words_ratio': {text_dist_words_ratio_range}")


Even though the 'mistakes_dist_ratio' and 'text_dist_words_ratio' features already fall within relatively small ranges, scaling them ensures consistency and comparability with the TF-IDF features. This normalization prevents any single feature from disproportionately influencing the model due to differing scales.

# Exclude 'mistakes_dist_ratio' and 'text_dist_words_ratio' from scaling
features_to_scale = [
    'reading_time',
    'mistakes_dist_ratio',
    'polysyllabcount',
    'sentence_count',
    'difficult_words',
    'comma_count',
    'transitional_phrases_c',
    'text_dist_words_ratio'
]

# Scale the selected numerical features
scaler = StandardScaler()
scaled_features = scaler.fit_transform(numerical_features_df[features_to_scale])

# Convert scaled numerical features back to DataFrame
scaled_features_df = pd.DataFrame(scaled_features, columns=features_to_scale)

# Include the unscaled features
#unscaled_features_df = numerical_features_df[['mistakes_dist_ratio', 'text_dist_words_ratio']]

# Check the minimum and maximum values for each feature
min_values = scaled_features_df.min()
max_values = scaled_features_df.max()

# Create a DataFrame to display the range of each feature
range_df = pd.DataFrame({'min': min_values, 'max': max_values})

# Display the range
print(range_df)

### Combine numerical scaled features with TF_IDF vectors

# Combine scaled and unscaled features
#combined_numerical_features_df = pd.concat([scaled_features_df, unscaled_features_df], axis=1)

# Combine the numerical features with the TF-IDF features
combined_features_df = pd.concat([scaled_features_df, tfidf_features_df], axis=1)

# Print the count of the combined features
print(f"Total number of features: {combined_features_df.shape[1]}")

# Export the combined features to a CSV file
combined_features_df.to_csv('combined_features.csv', index=False)
print('File exported')
